+++ apps/nextjs-example/src/components/transactionFlows/MultiAgent.tsx
@@ -45,6 +45,7 @@ export function MultiAgent() {
     await aptosClient(network).fundAccount({
       accountAddress: secondarySigner.accountAddress.toString(),
       amount: 100_000_000,
+      options: { waitForIndexer: false }
     });
     setSecondarySignerAccount(secondarySigner);
 
@@ -86,16 +87,16 @@ export function MultiAgent() {
   };
 
   const onSubmitTransaction = async () => {
-    if (!transactionToSubmit) {
-      throw new Error("No Transaction to sign");
-    }
-    if (!senderAuthenticator) {
-      throw new Error("No senderAuthenticator");
-    }
-    if (!secondarySignerAuthenticator) {
-      throw new Error("No secondarySignerAuthenticator");
-    }
     try {
+      if (!transactionToSubmit) {
+        throw new Error("No Transaction to sign");
+      }
+      if (!senderAuthenticator) {
+        throw new Error("No senderAuthenticator");
+      }
+      if (!secondarySignerAuthenticator) {
+        throw new Error("No secondarySignerAuthenticator");
+      }
       const response = await submitTransaction({
         transaction: transactionToSubmit,
         senderAuthenticator: senderAuthenticator,
@@ -106,6 +107,11 @@ export function MultiAgent() {
         description: <TransactionHash hash={response.hash} network={network} />,
       });
     } catch (error) {
+      toast({
+        variant: "destructive",
+        title: "Error",
+        description: "Unable to submit multiagent Transaction.",
+      });
       console.error(error);
     }
   };