+++ packages/wallet-adapter-core/src/AIP62StandardWallets/sdkWallets.ts
@@ -1,15 +1,35 @@
-import { DevTWallet, TWallet } from "@atomrigslab/aptos-wallet-adapter";
 import { AptosConnectWallet } from "@aptos-connect/wallet-adapter-plugin";
-import { AptosStandardWallet } from "./WalletStandard";
 import { Network } from "@aptos-labs/ts-sdk";
+import { DevTWallet, TWallet } from "@atomrigslab/aptos-wallet-adapter";
+import { MizuWallet } from "@mizuwallet-sdk/aptos-wallet-adapter";
 import { DappConfig } from "../WalletCore";
+import { AptosStandardWallet } from "./WalletStandard";
 
 export function getSDKWallets(dappConfig?: DappConfig) {
   const sdkWallets: AptosStandardWallet[] = [];
 
   // Need to check window is defined for AptosConnect
   if (typeof window !== "undefined") {
-    sdkWallets.push(new AptosConnectWallet({ network: dappConfig?.network }));
+    sdkWallets.push(
+      new AptosConnectWallet({
+        network: dappConfig?.network,
+        dappId: dappConfig?.aptosConnectDappId,
+      })
+    );
+
+    if (
+      dappConfig?.mizuwallet &&
+      dappConfig?.network &&
+      [Network.MAINNET, Network.TESTNET].includes(dappConfig.network)
+    ) {
+      sdkWallets.push(
+        new MizuWallet({
+          network: dappConfig.network as any,
+          manifestURL: dappConfig.mizuwallet.manifestURL,
+          appId: dappConfig.mizuwallet.appId,
+        }) as any
+      );
+    }
   }
 
   // Push production wallet if env is production, otherwise use dev wallet
@@ -20,6 +40,9 @@ export function getSDKWallets(dappConfig?: DappConfig) {
   } else {
     sdkWallets.push(new DevTWallet() as any);
   }
+
+  // Add new SDK wallet plugins (ones that should be installed as packages) here:
+  // Ex. sdkWallets.push(new YourSDKWallet(dappConfig))
 
   return sdkWallets;
 }