+++ packages/wallet-adapter-core/src/WalletCore.ts
@@ -87,6 +87,10 @@ export type IAptosWallet = AptosStandardWallet & Wallet;
 export interface DappConfig {
   network: Network;
   aptosConnectDappId?: string;
+  mizuwallet?: {
+    manifestURL: string;
+    appId?: string;
+  };
 }
 
 /** Any wallet that can be handled by `WalletCore`.
@@ -149,13 +153,15 @@ export class WalletCore extends EventEmitter<WalletCoreEvents> {
   constructor(
     plugins: ReadonlyArray<Wallet>,
     optInWallets: ReadonlyArray<AvailableWallets>,
-    dappConfig?: { network: Network }
+    dappConfig?: DappConfig
   ) {
     super();
+
     this._wallets = plugins;
     this._optInWallets = optInWallets;
     this._dappConfig = dappConfig;
     this._sdkWallets = getSDKWallets(this._dappConfig);
+
     // Strategy to detect AIP-62 standard compatible extension wallets
     this.fetchExtensionAIP62AptosWallets();
     // Strategy to detect AIP-62 standard compatible SDK wallets.
@@ -170,6 +176,27 @@ export class WalletCore extends EventEmitter<WalletCoreEvents> {
 
   private scopePollingDetectionStrategy() {
     this._wallets?.forEach((wallet: Wallet) => {
+      // Hot fix to manage Pontem wallet to not show duplications if user has the
+      // new standard version installed. Pontem uses "Pontem" wallet name for previous versions and
+      // "Pontem Wallet" with new version
+      const existingStandardPontemWallet = this._standard_wallets.find(
+        (wallet) => wallet.name == "Pontem Wallet"
+      );
+      if (wallet.name === "Pontem" && existingStandardPontemWallet) {
+        return;
+      }
+
+      /**
+       * Manage potential duplications when a plugin is installed in a dapp
+       * but the existing wallet installed is on AIP-62 new standard - in that case we dont want to
+       * include the plugin wallet (i.e npm package)
+       */
+      const existingWalletIndex = this._standard_wallets.findIndex(
+        (standardWallet) => standardWallet.name == wallet.name
+      );
+      if (existingWalletIndex !== -1) return;
+
+      // push the plugin wallet to the all_wallets array
       this._all_wallets.push(wallet);
       if (!wallet.readyState) {
         wallet.readyState =
@@ -227,6 +254,19 @@ export class WalletCore extends EventEmitter<WalletCoreEvents> {
       // as a not detected wallet
       if (existingPluginWalletIndex !== -1) return;
 
+      // Hot fix to manage Pontem wallet to not show duplications if user has the
+      // new standard version installed. Pontem uses "Pontem" wallet name for previous versions and
+      // "Pontem Wallet" with new version
+      const existingStandardPontemWallet = this.wallets.find(
+        (wallet) => wallet.name == "Pontem"
+      );
+      if (
+        supportedWallet.name === "Pontem Wallet" &&
+        existingStandardPontemWallet
+      ) {
+        return;
+      }
+
       // Check if we already have this wallet as a AIP-62 wallet standard
       const existingStandardWallet = this._standard_wallets.find(
         (wallet) => wallet.name == supportedWallet.name
@@ -283,8 +323,6 @@ export class WalletCore extends EventEmitter<WalletCoreEvents> {
    * @returns
    */
   excludeWallet(wallet: AptosStandardWallet): boolean {
-    // for now, we always include AptosConnect
-    if (isAptosConnectWallet(wallet)) return false;
     // If _optInWallets is not empty, and does not include the provided wallet,
     // return true to exclude the wallet, otherwise return false
     if (